net: shell: Add ipv6 command
Move "ipv6" command to a separate file.
